feat: split signal and byteorder into top-level modules
