+++ Kbuild
@@ -64,4 +64,6 @@ obj-$(CONFIG_LUNATIK_CRYPTO_RNG) += lib/luacrypto_rng.o
 obj-$(CONFIG_LUNATIK_CRYPTO_COMP) += lib/luacrypto_comp.o
 obj-$(CONFIG_LUNATIK_CPU) += lib/luacpu.o
 obj-$(CONFIG_LUNATIK_HID) += lib/luahid.o
+obj-$(CONFIG_LUNATIK_SIGNAL) += lib/luasignal.o
+obj-$(CONFIG_LUNATIK_BYTEORDER) += lib/luabyteorder.o
